Add dispute to communities
* Add new env var DISPUTE_CHANNEL to be used for disputes on main channel (not communities)
* Disabled '===' on eslint
* Created dispute module
* Refactoring messages.js
  * Moved getDetailedOrder() to util.js
  * Moved messages to dispute module
* Add locale messages related to disputes
