+++ .env-sample
@@ -60,3 +60,6 @@ MAX_ROUTING_FEE=0.001
 
 # Attempts to pay the invoice again when the payment failed
 PAYMENT_ATTEMPTS=2
+
+# Here will go the disputes from non community orders
+DISPUTE_CHANNEL='@p2plnbotDispute'
